+++ fractal/components/stickies/StickyBar/StickyBar.config.js
@@ -1,0 +1,9 @@
+module.exports = {
+    "title": "Sticky Bar",
+    "status": "ready",
+    "collated": true,
+    "preview": "@stickies-preview",
+    "context": {
+        "children": "Here is some text"
+    }
+};