+++ CHANGELOG.md
@@ -85,6 +85,13 @@ Updated styles to use a base-size-unit variable as design has changed from 10px 
 
 -----
 
+6/23/2017
+
+- Added sticky bar component
+
+
+-----
+
 6/24/2017
 
 - Updated some breakpoint functions and mix-ins